fix(opencode-orchestrator-mcp): tighten command transport supervision tests
